+++ src/main/formats/HOSA/HOSAScanner.cpp
@@ -62,7 +62,7 @@ void HOSAScanner::Scan(RawFile *file, void *info) {
 HOSASeq *HOSAScanner::SearchForHOSASeq(RawFile *file) {
   std::string name = file->tag.HasTitle() ? file->tag.title : removeExtFromPath(file->name());
 
-  uint32_t nFileLength = file->size();
+  size_t nFileLength = file->size();
   for (uint32_t i = 0; i + 4 < nFileLength; i++) {
     // Signature must match
     if (file->GetWordBE(i) != 0x484F5341 || file->GetByte(i + 4) != 'V')  //"HOSAV"
@@ -93,16 +93,16 @@ HOSASeq *HOSAScanner::SearchForHOSASeq(RawFile *file) {
 #define MIN_NUM_SAMPLES_COMPARE 5
 #define MIN_SAMPLES_MATCH 4
 HOSAInstrSet *HOSAScanner::SearchForHOSAInstrSet(RawFile *file, const PSXSampColl *sampcoll) {
-  size_t numSamples = sampcoll->samples.size();
+  int numSamples = static_cast<int>(sampcoll->samples.size());
   if (numSamples < MIN_NUM_SAMPLES_COMPARE) {
     return nullptr;
   }
 
   uint32_t *sampOffsets = new uint32_t[numSamples];
-  for (unsigned int i = 0; i < numSamples; i++)
+  for (int i = 0; i < numSamples; i++)
     sampOffsets[i] = sampcoll->samples[i]->dwOffset - sampcoll->dwOffset;
 
-  uint32_t nFileLength = file->size();
+  size_t nFileLength = file->size();
   for (uint32_t i = 0x20; i + 0x14 < nFileLength; i++) {
     if (RecursiveRgnCompare(file, i, 0, numSamples, 0, sampOffsets)) {
       for (; i >= 0x20; i -= 4) {
