+++ src/main/util/decompression.h
@@ -27,7 +27,7 @@ std::vector<char> zdecompress(T src) {
     throw std::runtime_error("Failed to init decompression");
   }
 
-  strm.avail_in = src.size();
+  strm.avail_in = static_cast<uint32_t>(src.size());
   strm.next_in = reinterpret_cast<z_const Bytef *>(src.data());
 
   unsigned actual_size = 0;
